feed entity 수정 - storeSize 타입 오작성 수정
- storeSize 오작성된 부분 수정(Int → Long)

+++ src/main/kotlin/com/bee/newsfeed_bee/feed/entity/Feed.kt
@@ -13,7 +13,7 @@ class Feed(
     var score: Int,
     var visitedDateTime: LocalDateTime,
     var waited: Boolean,
-    var storeSize: Long, // enum
+    var storeSize: Int, // enum
     var content: String,
 ) {
